requirements.txt
Flask==2.2.5
tensorflow==2.13.0
keras==2.13.1
numpy==1.23.5
Pillow==9.5.0
